+++ front/lib/models/assistant/actions/mcp.ts
@@ -197,8 +197,6 @@ export class AgentMCPAction extends WorkspaceAwareModel<AgentMCPAction> {
 
   declare outputItems: NonAttribute<AgentMCPActionOutputItem[]>;
   declare agentMessage?: NonAttribute<AgentMessage>;
-  // TODO(2025-07-23 aubin): remove this once the analytics have been moved to the resource.
-  declare stepContent?: NonAttribute<AgentStepContentModel>;
 }
 
 AgentMCPAction.init(
