+++ front/lib/resources/agent_step_content_resource.ts
@@ -5,6 +5,7 @@ import type {
   CreationAttributes,
   ModelStatic,
   Transaction,
+  WhereOptions,
 } from "sequelize";
 import { Op } from "sequelize";
 
@@ -16,7 +17,11 @@ import {
   AgentMCPActionOutputItem,
 } from "@app/lib/models/assistant/actions/mcp";
 import { AgentStepContentModel } from "@app/lib/models/assistant/agent_step_content";
-import { AgentMessage } from "@app/lib/models/assistant/conversation";
+import {
+  AgentMessage,
+  ConversationModel,
+  Message,
+} from "@app/lib/models/assistant/conversation";
 import { BaseResource } from "@app/lib/resources/base_resource";
 import { FileResource } from "@app/lib/resources/file_resource";
 import { frontSequelize } from "@app/lib/resources/storage";
@@ -28,6 +33,30 @@ import type { ModelId, Result } from "@app/types";
 import { removeNulls } from "@app/types";
 import { Err, Ok } from "@app/types";
 import type { AgentStepContentType } from "@app/types/assistant/agent_message_content";
+
+// AgentMCPAction serialized for analytics purposes.
+export type AnalyticsMCPAction = {
+  sId: string;
+  createdAt: string;
+  functionCallName: string | null;
+  params: Record<string, unknown>;
+  executionState: string;
+  isError: boolean;
+  conversationId: string;
+  messageId: string;
+};
+
+export type GetMCPActionsResult = {
+  actions: AnalyticsMCPAction[];
+  nextCursor: string | null;
+  totalCount: number;
+};
+
+type GetMCPActionsOptions = {
+  agentConfigurationId: string;
+  limit: number;
+  cursor?: string;
+};
 
 // Attributes are marked as read-only to reflect the stateless nature of our Resource.
 // eslint-disable-next-line @typescript-eslint/no-empty-interface, @typescript-eslint/no-unsafe-declaration-merging
@@ -464,4 +493,164 @@ export class AgentStepContentResource extends BaseResource<AgentStepContentModel
       );
     });
   }
+
+  static async getMCPActionsForAgent(
+    auth: Authenticator,
+    { agentConfigurationId, limit, cursor }: GetMCPActionsOptions
+  ): Promise<Result<GetMCPActionsResult, Error>> {
+    const owner = auth.getNonNullableWorkspace();
+
+    const whereClause: WhereOptions<AgentStepContentModel> = {
+      workspaceId: owner.id,
+      type: "function_call",
+    };
+
+    if (cursor) {
+      const cursorDate = new Date(cursor);
+      if (isNaN(cursorDate.getTime())) {
+        return new Err(new Error("Invalid cursor format"));
+      }
+      whereClause.createdAt = {
+        [Op.lt]: cursorDate,
+      };
+    }
+
+    try {
+      // Get total count for pagination
+      const totalCount = await AgentStepContentModel.count({
+        include: [
+          {
+            model: AgentMessage,
+            as: "agentMessage",
+            required: true,
+            where: {
+              agentConfigurationId: agentConfigurationId,
+            },
+            include: [
+              {
+                model: Message,
+                as: "message",
+                required: true,
+                include: [
+                  {
+                    model: ConversationModel,
+                    as: "conversation",
+                    required: true,
+                    where: {
+                      visibility: { [Op.ne]: "deleted" },
+                    },
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            model: AgentMCPAction,
+            as: "agentMCPActions",
+            required: true,
+          },
+        ],
+        where: whereClause,
+      });
+
+      // Get all function call step contents with their MCP actions
+      const stepContents = await AgentStepContentModel.findAll({
+        include: [
+          {
+            model: AgentMessage,
+            as: "agentMessage",
+            required: true,
+            where: {
+              agentConfigurationId: agentConfigurationId,
+            },
+            include: [
+              {
+                model: Message,
+                as: "message",
+                required: true,
+                include: [
+                  {
+                    model: ConversationModel,
+                    as: "conversation",
+                    required: true,
+                    where: {
+                      visibility: { [Op.ne]: "deleted" },
+                    },
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            model: AgentMCPAction,
+            as: "agentMCPActions",
+            required: true,
+          },
+        ],
+        where: whereClause,
+        order: [["createdAt", "DESC"]],
+        limit: limit + 1, // Fetch one extra to determine if there are more results
+      });
+
+      // Determine if there are more results and get the actual results
+      const hasMore = stepContents.length > limit;
+      const actualStepContents = hasMore ? stepContents.slice(0, limit) : stepContents;
+      const nextCursor = hasMore
+        ? actualStepContents[actualStepContents.length - 1].createdAt.toISOString()
+        : null;
+
+      // Flatten MCP actions from step contents
+      const actionsData: AnalyticsMCPAction[] = [];
+      for (const stepContent of actualStepContents) {
+        if (stepContent.agentMCPActions) {
+          for (const action of stepContent.agentMCPActions) {
+            actionsData.push(serializeMCPActionFromStepContent(stepContent, action));
+          }
+        }
+      }
+
+      return new Ok({
+        actions: actionsData,
+        nextCursor,
+        totalCount,
+      });
+    } catch (error) {
+      logger.error(
+        {
+          workspaceId: owner.id,
+          agentConfigurationId,
+          error,
+        },
+        "Failed to fetch MCP actions from database"
+      );
+      return new Err(new Error("Failed to fetch MCP actions from database"));
+    }
+  }
+}
+
+function serializeMCPActionFromStepContent(
+  stepContent: AgentStepContentModel,
+  action: AgentMCPAction
+): AnalyticsMCPAction {
+  assert(stepContent.agentMessage, "Agent message must exist");
+  assert(stepContent.agentMessage.message, "Message must exist");
+  assert(stepContent.agentMessage.message.conversation, "Conversation must exist");
+  assert(
+    stepContent.value.type === "function_call",
+    "Step content must be a function call"
+  );
+
+  return {
+    sId: MCPActionType.modelIdToSId({
+      id: action.id,
+      workspaceId: action.workspaceId,
+    }),
+    createdAt: action.createdAt.toISOString(),
+    functionCallName: stepContent.value.value.name,
+    params: JSON.parse(stepContent.value.value.arguments),
+    executionState: action.executionState,
+    isError: action.isError,
+    conversationId: stepContent.agentMessage.message.conversation.sId,
+    messageId: stepContent.agentMessage.message.sId,
+  };
 }
