+++ front/pages/api/w/[wId]/labs/mcp_actions/[agentId]/index.ts
@@ -4,12 +4,12 @@ import * as reporter from "io-ts-reporters";
 import { NumberFromString, withFallback } from "io-ts-types";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import type { GetMCPActionsResult } from "@app/lib/api/actions/agent_mcp_action_analytics";
-import { AgentMCPActionsAnalytics } from "@app/lib/api/actions/agent_mcp_action_analytics";
 import { getAgentConfiguration } from "@app/lib/api/assistant/configuration";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { Authenticator } from "@app/lib/auth";
 import { getFeatureFlags } from "@app/lib/auth";
+import type { GetMCPActionsResult } from "@app/lib/resources/agent_step_content_resource";
+import { AgentStepContentResource } from "@app/lib/resources/agent_step_content_resource";
 import { apiError } from "@app/logger/withlogging";
 import type { WithAPIErrorResponse } from "@app/types";
 
@@ -93,7 +93,7 @@ async function handler(
         });
       }
 
-      const result = await AgentMCPActionsAnalytics.getMCPActionsForAgent(
+      const result = await AgentStepContentResource.getMCPActionsForAgent(
         auth,
         {
           agentConfigurationId: agentConfiguration.sId,
